+++ CHANGES.md
@@ -1,3 +1,11 @@
+## [HEAD]
+
+- Deprecate usage of `<base href>` as an automatic substitute for `basename`
+  in `useBasename`. They have subtly different semantics ([#94])
+
+[HEAD]: https://github.com/mjackson/history/compare/latest...HEAD
+[#94]: https://github.com/mjackson/history/issues/94
+
 ## [v2.0.2]
 > Apr 14, 2016